Added testing framework and corrected require issue
There was an initial issue with requiring the 'ie' module,
which with component works properly, but is published
under a different name in npm ('component-ie'). Added
a try/catch to the head of the library to ensure
proper inclusion.

Testing is done via mocha, tests are located in the
`test/` folder, and can be run by using `npm test`. This
will chack all non-test scripts with feross/standard syntax
tester, then run mocha through istanbul, which  not only
runs tests, but also gives coverage reporting. You can access
the coverage report in `coverage/lcov-report/index.html`

To ensure that the feross/standard passed without issue, I made
minor alterations to the styling of `index.js` (space before function
parentheses, removed semicolons), as well as added a function name
to the main `module.exports` function, as it was anonymous previously.